feat(openapi): promote OpenAPI 3.0.3 helpers to public socle::openapi module
Adds src/openapi.rs with the five items required by issue #13:
- merge_health_paths(&mut OpenApi, &str) — idempotent health path injection
- to_3_0_pretty_json(&OpenApi) -> Result<String> — 3.0.3 serialisation
- rewrite_nullable_for_progenitor(&mut Value) — nullable array rewriting
- BearerAuthAddon — utoipa::Modify adding securitySchemes.BearerAuth
- generate_openapi_binary! macro — updated to use new &mut signatures

Removes old inline openapi module from lib.rs and cleans up dead adapter
functions that were superseded by the new public module.

Closes #13
