+++ src/lib.rs
@@ -107,8 +107,9 @@ macro_rules! generate_openapi_binary {
         fn main() {
             use utoipa::OpenApi as _;
             use $crate::openapi::{merge_health_paths, to_3_0_pretty_json};
-            let api = merge_health_paths(<$api_doc>::openapi(), $health_path);
-            match to_3_0_pretty_json(api) {
+            let mut doc = <$api_doc>::openapi();
+            merge_health_paths(&mut doc, $health_path);
+            match to_3_0_pretty_json(&doc) {
                 Ok(json) => println!("{json}"),
                 Err(e) => {
                     eprintln!("generate-openapi: serialize failed: {e}");
@@ -119,51 +120,9 @@ macro_rules! generate_openapi_binary {
     };
 }
 
-/// OpenAPI helpers for callers generating specs outside of [`ServiceBootstrap::serve`].
+/// OpenAPI 3.0.3 helpers for `axum` + `utoipa` + `progenitor` consumers.
 #[cfg(feature = "openapi")]
-pub mod openapi {
-    /// Merge socle's built-in health path definitions into `api`.
-    pub fn merge_health_paths(
-        api: utoipa::openapi::OpenApi,
-        health_path: &str,
-    ) -> utoipa::openapi::OpenApi {
-        crate::adapters::openapi::merge_health_paths(api, health_path)
-    }
-
-    /// Rewrite OpenAPI 3.1 nullable type arrays to OpenAPI 3.0 `nullable: true` form.
-    pub fn rewrite_nullable_for_progenitor(json: String) -> String {
-        crate::adapters::openapi::rewrite_nullable_for_progenitor(json)
-    }
-
-    /// Serialize `api` as a valid OpenAPI **3.0.3** pretty-printed JSON string.
-    pub fn to_3_0_pretty_json(api: utoipa::openapi::OpenApi) -> serde_json::Result<String> {
-        crate::adapters::openapi::to_3_0_pretty_json(api)
-    }
-
-    /// Strip `content` from non-2xx response objects in a serialised OpenAPI value.
-    ///
-    /// Call this explicitly after [`to_3_0_pretty_json`] if desired. This is
-    /// opt-in — [`to_3_0_pretty_json`] does **not** call it automatically.
-    pub fn strip_non_success_response_content(val: &mut serde_json::Value) {
-        crate::adapters::openapi::strip_non_success_response_content(val)
-    }
-
-    pub use crate::generate_openapi_binary;
-
-    /// utoipa [`utoipa::Modify`] plugin that registers a `bearerAuth` HTTP Bearer security scheme.
-    pub struct BearerAuthAddon;
-
-    impl utoipa::Modify for BearerAuthAddon {
-        fn modify(&self, openapi: &mut utoipa::openapi::OpenApi) {
-            use utoipa::openapi::security::{Http, HttpAuthScheme, SecurityScheme};
-            let components = openapi.components.get_or_insert_with(Default::default);
-            components.add_security_scheme(
-                "bearerAuth",
-                SecurityScheme::Http(Http::new(HttpAuthScheme::Bearer)),
-            );
-        }
-    }
-}
+pub mod openapi;
 
 /// Re-exports of the underlying crates.
 pub mod reexports {
